+++ src/connectedk8s/azext_connectedk8s/custom.py
@@ -53,7 +53,7 @@ from kubernetes import config
 from kubernetes.config.kube_config import KubeConfigMerger
 from packaging import version
 
-import azext_connectedk8s._clientproxyutils as clientproxyutils
+import azext_connectedk8s.clientproxyhelper._utils as clientproxyutils
 import azext_connectedk8s._constants as consts
 import azext_connectedk8s._precheckutils as precheckutils
 import azext_connectedk8s._troubleshootutils as troubleshootutils
